gui/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
websockets>=13.0
python-dotenv>=1.0.1
pydantic>=2.0.0
PyYAML>=6.0.1
openai>=1.0.0
anthropic>=0.40.0
google-genai>=1.0.0
sse-starlette>=2.0.0
httpx>=0.25.0
